renamed data jobs with first-time names to be generic
